requirements.txt
flask
opencv-python-headless
pillow
numpy
gunicorn